+++ charts/ccsm-helm/test/identity/goldenfiles_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package operate
+package identity
 
 import (
 	"path/filepath"
